Update with new version of the JSON library

+++ include/gigamonkey/stratum/message_id.hpp
@@ -14,27 +14,27 @@ namespace Gigamonkey::Stratum {
     // Message id is either an integer or string, must be a unique identifier. 
     struct message_id : JSON {
         
-        message_id(const uint64& d) : JSON(d) {}
-        message_id(const string& x) : JSON(x) {}
+        message_id (const uint32 &d) : JSON (d) {}
+        message_id (const string &x) : JSON (x) {}
         
-        bool valid() const;
+        bool valid () const;
         
-        static bool valid(const JSON&); 
+        static bool valid (const JSON &); 
         
     private:
-        message_id() : JSON{} {}
-        message_id(const JSON& j) : JSON(j) {}
+        message_id () : JSON {} {}
+        message_id (const JSON &j) : JSON (j) {}
         
         friend struct request;
         friend struct response;
     };
     
-    bool inline message_id::valid(const JSON& j) {
-        return j.is_number_unsigned() || j.is_string();
+    bool inline message_id::valid (const JSON &j) {
+        return j.is_number_unsigned () || j.is_string ();
     }
     
-    bool inline message_id::valid() const {
-        return valid(*this);
+    bool inline message_id::valid () const {
+        return valid (*this);
     }
     
 }
